fix(components): update the path of Cuboidform
